Move deprecated function calls in api testers to deprecated files
