+++ api-tester/src/main/java/com/revenuecat/apitester/java/DeprecatedPurchasesAPI.java
@@ -2,11 +2,20 @@ package com.revenuecat.apitester.java;
 
 import android.app.Activity;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import com.android.billingclient.api.SkuDetails;
+import com.revenuecat.purchases.CustomerInfo;
 import com.revenuecat.purchases.Package;
 import com.revenuecat.purchases.Purchases;
+import com.revenuecat.purchases.PurchasesError;
 import com.revenuecat.purchases.UpgradeInfo;
+import com.revenuecat.purchases.interfaces.ProductChangeCallback;
+import com.revenuecat.purchases.interfaces.PurchaseCallback;
 import com.revenuecat.purchases.models.StoreProduct;
+import com.revenuecat.purchases.models.StoreTransaction;
+import com.revenuecat.purchases.models.SubscriptionOption;
 
 @SuppressWarnings({"unused"})
 final class DeprecatedPurchasesAPI {
@@ -15,8 +24,35 @@ final class DeprecatedPurchasesAPI {
                       final SkuDetails skuDetails,
                       final StoreProduct storeProduct,
                       final Package packageToPurchase,
+                      final SubscriptionOption subscriptionOption,
                       final UpgradeInfo upgradeInfo) {
+        final ProductChangeCallback purchaseChangeListener = new ProductChangeCallback() {
+            @Override
+            public void onCompleted(@Nullable StoreTransaction storeTransaction, @NonNull CustomerInfo customerInfo) {
+            }
+
+            @Override
+            public void onError(@NonNull PurchasesError error, boolean userCancelled) {
+            }
+        };
+        final PurchaseCallback makePurchaseListener = new PurchaseCallback() {
+            @Override
+            public void onCompleted(@NonNull StoreTransaction storeTransaction, @NonNull CustomerInfo customerInfo) {
+            }
+
+            @Override
+            public void onError(@NonNull PurchasesError error, boolean userCancelled) {
+            }
+        };
+
         purchases.setAllowSharingPlayStoreAccount(true);
+        Purchases.setDebugLogsEnabled(false);
+        purchases.purchaseProduct(activity, storeProduct, upgradeInfo, purchaseChangeListener);
+        purchases.purchaseProduct(activity, storeProduct, makePurchaseListener);
+        purchases.purchasePackage(activity, packageToPurchase, upgradeInfo, purchaseChangeListener);
+        purchases.purchasePackage(activity, packageToPurchase, makePurchaseListener);
+        purchases.purchaseSubscriptionOption(activity, subscriptionOption, upgradeInfo, purchaseChangeListener);
+        purchases.purchaseSubscriptionOption(activity, subscriptionOption, makePurchaseListener);
     }
 
 }
